# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/commune/topics": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "list_topics",
        "responses": {
          "200": {
            "description": "Topics retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "create_topic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Topic created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}